DataProcessors
To resolve the ScienceData/Part coherency in the model, it's now using full DataPages.  DataPages now implement IEquatable, so they can be *value* compared, instead of ref compared.

the controller just keeps the selected list.
to that effect, m_sciDataParts is gone(in a dp anyway)
models m_radioModules got changed to the ModuleRadioTransmitter because non virt functions were needed.

changes propagate through the controller and view to match the model having the DataPages, but it was actually pretty painless change.

DataProcessors are an abstraction on a common pattern in altering model sciencedata: for each datum we'd do a 1 shot step, poll on something a while, then do a final step after, then move onto datum i+1.  The DataProcessor automates this, using Action or Function hooks to let the model treat each differently.  Hooks can be null for skipping steps (except the poll2).

The discard & move actions cheat the step1 of the pattern, by doing all discards at once.  If we did them in series, it would take n frames, instead of 1 as is.  Polling on the removes fixes the sync problems we had before with this.

The lab copies fire coroutines off of the controller (must use GO), and listen for the finish callback to add the Research to the lab.

The transmits temporarily subscribe to the scienceRecieve event, as the transmitter leaves us no othe rway to listen for when it's done transmitting, and before it retracts again.